Add settings to turn registration and invite codes on/off.
