+++ saylua/modules/users/urls.py
@@ -4,11 +4,11 @@ from . import views
 urlpatterns = [
     # Login URLs
     url('/login/', view_func=views.login.login, name='login', methods=['GET', 'POST']),
-    url('/login/recover/', view_func=views.login.recover_login, name='login_recover',
+    url('/login/recover/', view_func=views.recover.recover_login, name='login_recover',
         methods=['GET', 'POST']),
-    url('/login/reset/<user>/<code>/', view_func=views.login.reset_password, name='login_reset'),
+    url('/login/reset/<user>/<code>/', view_func=views.recover.reset_password, name='login_reset'),
     url('/logout/', view_func=views.login.logout, name='logout'),
-    url('/register/', view_func=views.login.register, name='register', methods=['GET', 'POST']),
+    url('/register/', view_func=views.register.register, name='register', methods=['GET', 'POST']),
 
     # Online users
     url('/online/', view_func=views.online.users_online, name='online'),
