+++ saylua/modules/users/views/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa
-from . import admin, login, online, profile, settings
+from . import admin, login, online, profile, recover, register, settings
